minor code fix
review 적용
