+++ Assets/Scripts/Common/ScoreResult.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6b03be2c21e5065498336d42a8c887a1
