+++ Assets/Scripts/Common/WinningScoreData.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ce66e58183217f04e9f64c25c06ca1f2
